refactor(web): Phase 7 — extract integrations glue to @openloomi/integrations-runtime
Move the 5 dep-free integrations glue files (authorization errors, oauth
callback script, task-integration inference, platform visuals, platform
connectability) out of apps/web into a new shared package
@openloomi/integrations-runtime. apps/web/lib/integrations/{...}.ts now
become re-export shims over the new package subpaths, matching the same
pattern used in earlier phases (cron, loop, insights).

apps/web keeps the heavy DB / session / token-manager glue for now;
it will move in a later phase once runtime services have their own
Hono/MCP daemons.

Build + typecheck + 2253 tests pass.
