+++ apps/web/package.json
@@ -57,6 +57,7 @@
 		"@openloomi/indexeddb": "workspace:*",
 		"@openloomi/insights": "workspace:*",
 		"@openloomi/integrations": "workspace:*",
+		"@openloomi/integrations-runtime": "workspace:*",
 		"@openloomi/loop": "workspace:*",
 		"@openloomi/memory-consolidation": "workspace:*",
 		"@openloomi/memory-store": "workspace:*",
